+++ tests/test_core/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for core modules / internals that don't really fit elsewhere."""
