+++ tests/test_core/test_error_handling.py
