+++ tests/test_core/test_metadata.py
