+++ tests/test_core/test_utils.py
